THIS IS THE FINAL VERSION OF REN'PY I'LL USE

+++ renpy/common/00updater.rpy
@@ -1,4 +1,4 @@
-﻿# Copyright 2004-2025 Tom Rothamel <pytom@bishoujo.us>
+﻿# Copyright 2004-2024 Tom Rothamel <pytom@bishoujo.us>
 #
 # Permission is hereby granted, free of charge, to any person
 # obtaining a copy of this software and associated documentation files
@@ -41,15 +41,16 @@ init -1500 python in updater:
     import zlib
     import codecs
     import io
+    import future.utils
 
     def urlopen(url):
         import requests
-        return io.BytesIO(requests.get(url, proxies=renpy.exports.proxies, timeout=15).content)
+        return io.BytesIO(requests.get(url).content)
 
     def urlretrieve(url, fn):
         import requests
 
-        data = requests.get(url, proxies=renpy.exports.proxies, timeout=15).content
+        data = requests.get(url).content
 
         with open(fn, "wb") as f:
             f.write(data)
@@ -68,6 +69,95 @@ init -1500 python in updater:
     # A map from update URL to the time we last checked that URL.
     if persistent._update_last_checked is None:
         persistent._update_last_checked = { }
+
+    # A file containing deferred update commands, one per line. Right now,
+    # there are two commands:
+    # R <path>
+    #     Rename <path>.new to <path>.
+    # D <path>
+    #     Delete <path>.
+    # Deferred commands that cannot be accomplished on start are ignored.
+    DEFERRED_UPDATE_FILE = os.path.join(config.renpy_base, "update", "deferred.txt")
+    DEFERRED_UPDATE_LOG = os.path.join(config.renpy_base, "update", "log.txt")
+
+    def process_deferred_line(l):
+
+        cmd, _, fn = l.partition(" ")
+
+        if cmd == "R":
+            if os.path.exists(fn + ".new"):
+
+                if os.path.exists(fn):
+                    os.unlink(fn)
+
+                os.rename(fn + ".new", fn)
+
+        elif cmd == "D":
+
+            if os.path.exists(fn):
+                os.unlink(fn)
+
+        elif cmd == "":
+
+            pass
+
+        else:
+            raise Exception("Bad command. %r (%r %r)" % (l, cmd, fn))
+
+    def process_deferred():
+        if not os.path.exists(DEFERRED_UPDATE_FILE):
+            return
+
+        # Give a previous process time to quit (and let go of the
+        # open files.)
+        time.sleep(3)
+
+        try:
+            log = open(DEFERRED_UPDATE_LOG, "a")
+        except Exception:
+            log = io.BytesIO()
+
+        with log:
+            with open(DEFERRED_UPDATE_FILE, "r") as f:
+                for l in f:
+
+                    l = l.rstrip("\r\n")
+
+                    log.write(l)
+
+                    try:
+                        process_deferred_line(l)
+                    except Exception:
+                        traceback.print_exc(file=log)
+
+            try:
+                os.unlink(DEFERRED_UPDATE_FILE + ".old")
+            except Exception:
+                pass
+
+            try:
+                os.rename(DEFERRED_UPDATE_FILE, DEFERRED_UPDATE_FILE + ".old")
+            except Exception:
+                traceback.print_exc(file=log)
+
+    # Process deferred updates on startup, if any exist.
+
+    process_deferred()
+
+    DELETED = os.path.join(config.renpy_base, "update", "deleted")
+
+    def process_deleted():
+        if not os.path.exists(DELETED):
+            return
+
+        import shutil
+
+        try:
+            shutil.rmtree(DELETED)
+        except Exception as e:
+            pass
+
+    process_deleted()
 
     def zsync_path(command):
         """
@@ -110,7 +200,7 @@ init -1500 python in updater:
             The state that the updater is in.
 
         self.message
-            In an error state, the error message that occurred.
+            In an error state, the error message that occured.
 
         self.progress
             If not None, a number between 0.0 and 1.0 giving some sort of
@@ -123,7 +213,7 @@ init -1500 python in updater:
 
         # Here are the possible states.
 
-        # An error occurred during the update process.
+        # An error occured during the update process.
         # self.message is set to the error message.
         ERROR = "ERROR"
 
@@ -335,7 +425,7 @@ init -1500 python in updater:
                     self.log.flush()
 
             except Exception as e:
-                self.message = _type(e).__name__ + ": " + str(e)
+                self.message = _type(e).__name__ + ": " + unicode(e)
                 self.can_cancel = False
                 self.can_proceed = True
                 self.state = self.ERROR
@@ -456,7 +546,7 @@ init -1500 python in updater:
             url = urlparse.urljoin(self.url, self.updates[module]["rpu_url"])
 
             try:
-                resp = requests.get(url, proxies=renpy.exports.proxies, timeout=15)
+                resp = requests.get(url)
                 resp.raise_for_status()
             except Exception as e:
                 raise UpdateError(__("Could not download file list: ") + str(e))
@@ -577,8 +667,6 @@ init -1500 python in updater:
 
             # 8. Finish up.
 
-            self.save_state()
-
             persistent._update_version[self.url] = None
 
             if self.restart:
@@ -681,8 +769,8 @@ init -1500 python in updater:
                 persistent._update_version[self.url] = None
                 return
 
-            self.pretty_version = build.version or build.directory_name
-            persistent._update_version[self.url] = self.pretty_version
+            pretty_version = build.version or build.directory_name
+            persistent._update_version[self.url] = pretty_version
 
             if self.check_only:
                 renpy.restart_interaction()
@@ -897,17 +985,11 @@ init -1500 python in updater:
             # Does updates.json need to be verified?
             require_verified = False
 
-            # The key for either the current game, or a downloaded game.
-            key = os.path.join(self.updatedir, "key.pem")
+            # New-style ECDSA signature.
+            key = os.path.join(config.basedir, "update", "key.pem")
 
-            # A key that comes with the downloader app, to seed it.
             if not os.path.exists(key):
-                key = os.path.join(config.basedir, "public_key.pem")
-
-            # This game's included key, used for initial downloades if the downloader app
-            # doesn't have a key.
-            if not os.path.exists(key):
-                key = os.path.join(config.basedir, "update", "key.pem")
+                key = os.path.join(self.updatedir, "key.pem")
 
             if os.path.exists(key):
                 require_verified = True
@@ -967,12 +1049,8 @@ init -1500 python in updater:
 
             self.updates = json.loads(updates_json)
 
-            if "RENPY_TEST_MONKEYPATCH" in os.environ:
-                with open(os.environ["RENPY_TEST_MONKEYPATCH"], "r") as f:
-                    monkeypatch = f.read()
-                    exec(monkeypatch, globals(), globals())
-            elif verified and "monkeypatch" in self.updates:
-                exec(self.updates["monkeypatch"], globals(), globals())
+            if verified and "monkeypatch" in self.updates:
+                future.utils.exec_(self.updates["monkeypatch"], globals(), globals())
 
         def add_dlc_state(self, name):
 
@@ -1340,7 +1418,7 @@ init -1500 python in updater:
             self.log.write("downloading %r\n" % url)
             self.log.flush()
 
-            resp = requests.get(url, stream=True, proxies=renpy.exports.proxies, timeout=15)
+            resp = requests.get(url, stream=True)
 
             if not resp.ok:
                 raise UpdateError(_("The update file was not downloaded."))
@@ -1487,12 +1565,20 @@ init -1500 python in updater:
 
             for path in self.moves:
 
+                self.unlink(path)
+
+                if os.path.exists(path):
+                    self.log.write("could not rename file %s" % path)
+
+                    with open(DEFERRED_UPDATE_FILE, "a") as f:
+                        f.write("R " + path + "\r\n")
+
+                    continue
+
                 try:
-                    self.unlink(path)
                     os.rename(path + ".new", path)
                 except Exception:
-                    renpy.update.deferred.defer_rename(path)
-
+                    pass
 
         def delete_obsolete(self):
             """
@@ -1529,13 +1615,15 @@ init -1500 python in updater:
                 self.unlink(i)
 
                 if os.path.exists(i):
-                    renpy.update.deferred.defer_unlink(i)
+                    self.log.write("could not delete file %s" % i)
+                    with open(DEFERRED_UPDATE_FILE, "a") as f:
+                        f.write("D " + i + "\r\n")
 
             for i in old_directories:
                 try:
                     os.rmdir(i)
                 except Exception:
-                    renpy.update.deferred.defer_unlink(i)
+                    pass
 
         def save_state(self):
             """
@@ -1545,7 +1633,7 @@ init -1500 python in updater:
             fn = os.path.join(self.updatedir, "current.json")
 
             with open(fn, "w") as f:
-                json.dump(self.new_state, f, indent=2)
+                json.dump(self.new_state, f)
 
         def clean(self, fn):
             """
@@ -1915,7 +2003,6 @@ init -1500 python in updater:
 init -1500:
 
     screen updater(u):
-        layer config.interface_layer
 
         add "#000"
 
@@ -1932,7 +2019,7 @@ init -1500:
                 vbox:
 
                     if u.state == u.ERROR:
-                        text _("An error has occurred:")
+                        text _("An error has occured:")
                     elif u.state == u.CHECKING:
                         text _("Checking for updates.")
                     elif u.state == u.UPDATE_NOT_AVAILABLE:
@@ -1974,7 +2061,6 @@ init -1500:
 
 
     screen downloader(u):
-        layer config.interface_layer
 
         style_prefix "downloader"
 
@@ -1989,7 +2075,7 @@ init -1500:
             elif u.state == u.FINISHING or u.state == u.DONE:
                 text _("The game data has been downloaded.")
             else: # An error or unknown state.
-                text _("An error occurred when trying to download game data:")
+                text _("An error occured when trying to download game data:")
 
                 if u.message is not None:
                     text "[u.message!q]"
